readd stateList, genesisContract, Startting parameters
Former-commit-id: 1aa2cbd180a9b4edbfa03004dc703b4c9ff577dc

+++ mix/qml/DeployContractStep.qml
@@ -490,7 +490,8 @@ Rectangle {
 					id: clearDeployAction
 					onTriggered: {
 						worker.forceStopPooling()
-						fileIo.deleteDir(projectModel.deploymentDir)
+						if (projectModel.deploymentDir && projectModel.deploymentDir !== "")
+							fileIo.deleteDir(projectModel.deploymentDir)
 						projectModel.cleanDeploymentStatus()
 						deploymentDialog.steps.reset()
 					}
